refactor: move GeneratePodPatch under util package
